More refactoring and renaming
