--- app/src/main/java/com/minjee/favoritepokemon/ui/favorites/allpokemon/AllPokemonTabViewModel.kt
@@ -1,16 +1,0 @@
-package com.minjee.favoritepokemon.ui.favorites.allpokemon
-
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import com.minjee.favoritepokemon.data.PokemonList
-import com.minjee.favoritepokemon.data.PokemonRepository
-import com.minjee.favoritepokemon.data.PokemonRequest
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-
-class AllPokemonTabViewModel(private val pokemonRepository: PokemonRepository) : ViewModel() {
-
-
-}
