+++ app/src/main/java/com/minjee/favoritepokemon/ui/favorites/firstTeam/FirstTeamTabViewModel.kt
@@ -1,0 +1,8 @@
+package com.minjee.favoritepokemon.ui.favorites.firstTeam
+
+import androidx.lifecycle.ViewModel
+import com.minjee.favoritepokemon.data.PokemonRepository
+
+class FirstTeamTabViewModel(private val pokemonRepository: PokemonRepository) : ViewModel() {
+
+}
