--- app/src/main/java/com/minjee/favoritepokemon/ui/favorites/mypokemon/MyPokemonTabViewModel.kt
@@ -1,8 +1,0 @@
-package com.minjee.favoritepokemon.ui.favorites.mypokemon
-
-import androidx.lifecycle.ViewModel
-import com.minjee.favoritepokemon.data.PokemonRepository
-
-class MyPokemonTabViewModel(private val pokemonRepository: PokemonRepository) : ViewModel() {
-
-}
