+++ app/src/main/java/com/minjee/favoritepokemon/ui/favorites/secondTeam/SecondTeamTabFragment.kt
@@ -1,18 +1,21 @@
-package com.minjee.favoritepokemon.ui.favorites.mypokemon
+package com.minjee.favoritepokemon.ui.favorites.secondTeam
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.minjee.favoritepokemon.databinding.FragmentTabMyPokemonBinding
+import com.minjee.favoritepokemon.databinding.FragmentTabTeamsBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class MyPokemonTabFragment: Fragment() {
+/*
+ *      Fragment which will display a team of Pokemon chosen by the user
+ */
+class SecondTeamTabFragment : Fragment() {
 
-   private val myPokemonViewModel: MyPokemonTabViewModel by viewModel()
+    private val viewModel: SecondTeamTabViewModel by viewModel()
 
-    private var _binding: FragmentTabMyPokemonBinding? = null
+    private var _binding: FragmentTabTeamsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -20,7 +23,7 @@ class MyPokemonTabFragment: Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentTabMyPokemonBinding.inflate(inflater, container, false)
+        _binding = FragmentTabTeamsBinding.inflate(inflater, container, false)
         return binding.root
     }
