+++ app/src/main/java/com/minjee/favoritepokemon/ui/favorites/secondTeam/SecondTeamTabViewModel.kt
@@ -1,0 +1,8 @@
+package com.minjee.favoritepokemon.ui.favorites.secondTeam
+
+import androidx.lifecycle.ViewModel
+import com.minjee.favoritepokemon.data.PokemonRepository
+
+class SecondTeamTabViewModel(private val pokemonRepository: PokemonRepository) : ViewModel() {
+
+}
