+++ app/src/main/res/layout/fragment_tab_teams.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@color/red"
     android:layout_height="match_parent"
-    tools:context=".ui.favorites.allpokemon.AllPokemonTabFragment">
+    tools:context=".ui.favorites.firstTeam.FirstTeamTabFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/allPokemonTabRecyclerview"
